Add upper bounds to FFTW and DSP Compats
FFTW v0-1.0 re-exported the AbstractFFTs interface without specifying an upper
bound on the version of AbstractFFTs. I have changed the Compat.toml for FFTW to
place a upper bound on AbstractFFTs, so that packages get a consistent interface
when using FFTW.

I have also changed the Compat.toml for DSP, to avoid breakage caused by
AbstractFFTs v0.5 taking some functions that were previously defined in DSP. In
addition, the Compat.toml for DSP showed versions 0-0.5 being compatible with
Julia 1, even though this was not the case according to the early versions of
REQUIRE or the Project.toml file, when it existed. Testing shows that DSP 0.5
errors on Julia 1, so I have removed this table from the Compat.toml in DSP.